backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-socketio==5.11.1
python-dotenv==1.0.1
pydantic-settings==2.2.1
aiofiles==23.2.1
python-multipart==0.0.9
httpx==0.27.0
